musyx inpInit takes the McmdVoiceState it initialises
The last musyx int-as-pointer parameter with a named type on the other end: the body
opened `McmdVoiceState* vs = (McmdVoiceState*)state;` and then tested `state != 0`, and
both call sites already hold the typed value -- `macHandleActive` passed `(u32)sv` and
`synthInit` passed a literal 0.  Every other declaration in inp_ctrl.h already takes
`McmdVoiceState *`.  The alias existed only to carry the cast, so it goes with it.

Byte-identical: 1006 objects vs a pristine control at this parent, 0 differ, 4
RENUMBERED (literal pool only) -- snd_midictrl.o, mcmd_exec.o and synth.o all unmoved.

+++ src/musyx/runtime/mcmd_exec.c
@@ -623,7 +623,7 @@ void macHandleActive(McmdVoiceState* sv)
         sv->keyGroup = 0;
         sv->vibModAddScale = 0;
         sv->treScale = 0;
-        inpInit((u32)sv);
+        inpInit(sv);
         lastNote = inpGetMidiLastNote(sv->midi, sv->midiSet);
         if (lastNote != 0xff)
         {
